#19192: Give up on time.xmlrpc.com as an xmlrpc network test.
time.xmlrpc.com has come and gone over the years, and has been gone again for
a while.  The test did test one thing that the current xmlrpc tests don't: the
use of multiple levels of attribute names in the call.  So in addition to
removing the network test, we add a test in xmlrpc of dotted name access.

There should also be a test for when dotted name access is disallowed, but
that requires more extensive test harness refactoring, and in any case was not
tested by the network test we are deleting, since it is a server-side setting.

This is a slightly simplified version of a patch by Vajrasky Kok.